refactor: split config validation and command execution into separate files
- Extract config-validation.ts from config.ts (validation logic)
- Extract command-executor.ts from message.ts (permission, context, execution)
- Reduces per-file complexity to satisfy Codacy's file-level analysis
- All functions CCN ≤ 7 (verified with lizard)
- tsc --noEmit passes clean
